use common-replica, rename package
