+++ suripu-queue/src/main/java/com/hello/suripu/queue/SuripuQueue.java
@@ -51,10 +51,10 @@ import com.hello.suripu.queue.models.QueueHealthCheck;
 import com.hello.suripu.queue.models.SenseDataDAO;
 import com.hello.suripu.queue.modules.RolloutQueueModule;
 import com.hello.suripu.queue.resources.StatsResource;
-import com.hello.suripu.queue.workers.TimelineQueueConsumerManager;
-import com.hello.suripu.queue.workers.TimelineQueueProcessor;
-import com.hello.suripu.queue.workers.TimelineQueueProducerManager;
-import com.hello.suripu.queue.workers.TimelineQueueWorkerCommand;
+import com.hello.suripu.queue.timeline.TimelineQueueConsumerManager;
+import com.hello.suripu.queue.timeline.TimelineQueueProcessor;
+import com.hello.suripu.queue.timeline.TimelineQueueProducerManager;
+import com.hello.suripu.queue.cli.TimelineQueueWorkerCommand;
 import com.yammer.dropwizard.Service;
 import com.yammer.dropwizard.config.Bootstrap;
 import com.yammer.dropwizard.config.Environment;
