+++ suripu-queue/src/main/java/com/hello/suripu/queue/timeline/TimelineGenerator.java
@@ -1,4 +1,4 @@
-package com.hello.suripu.queue.workers;
+package com.hello.suripu.queue.timeline;
 
 import com.google.common.base.Optional;
 import com.hello.suripu.core.models.Timeline;
