+++ suripu-queue/suripu-queue.prod.yml
@@ -10,9 +10,11 @@ graphite:
 
 common_db:
   driverClass: org.postgresql.Driver
+
   user: common
   password: hello-common
-  url: jdbc:postgresql://common.cdawj8qazvva.us-east-1.rds.amazonaws.com:5432/common
+  url: jdbc:postgresql://common-replica-1.cdawj8qazvva.us-east-1.rds.amazonaws.com:5432/common
+
   properties:
     hibernate.dialect: org.hibernate.spatial.dialect.postgis.PostgisDialect
